Day 12 - refactor
